Get rid of maps to improve performance
maps are lazy. they make it easy for the dev, but hard
for the machine. All of my code can essentially be
deterministic, because it's contained to a grid. Instead of
using a map to store "coords" let's start using slices,
because indexing into slices is _much_ easier for a
machine that computing hashes of maps.